+++ src/main/java/com/cleversoap/jota/interfaces/IRenderable.java
@@ -3,5 +3,5 @@ package com.cleversoap.jota.interfaces;
 public interface IRenderable
 {
     // TODO: See if delta is needed
-    void render(int delta);
+    void render(double delta);
 }
