+++ src/main/java/com/cleversoap/jota/scene/Scene.java
@@ -9,7 +9,7 @@ public class Scene
     {
     }
 
-    public void render(int delta)
+    public void render(double delta)
     {
     }
